create classes of todo-list/singleton, add types and adjust main.ts
