+++ src/main.ts
@@ -1,67 +1,17 @@
+import { todoList } from "./components/todo-list";
+
 const formElement = document.getElementById('todo-form')! as HTMLFormElement;
 
-type Todo = {
-    id: string
-    text: string
-    completed: boolean
-}
-
-let todos: Todo[] = []
 
 function handleFormSubmit(e: Event){
     e.preventDefault()
     const formData = new FormData(formElement)
     const todoValue = formData.get("todo-text") as string
 
-    const newTodo: Todo = {
-        id: crypto.randomUUID(),
-        text: todoValue,
-        completed: false
-    }
-    todos.push(newTodo)
-    renderTodos()
+    todoList.addTodo(todoValue)
+
     formElement.reset()
-    console.log(todos)
 }
 
 formElement.addEventListener('submit',handleFormSubmit)
 
-function renderTodos() {
-    const todoListElement = document.getElementById('todo-list') as HTMLUListElement
-    todoListElement.innerHTML = ''
-
-    todos.forEach((item) => {
-        const todoItemElement = document.createElement('li')
-        const todoSpanElement = document.createElement("span")
-        todoSpanElement.innerHTML = item.text
-
-        if (item.completed) {
-            todoItemElement.classList.add('completed')
-        }
-
-        todoItemElement.addEventListener('click', () => {
-            todos = todos.map((todo)=> {
-                if (todo.id === item.id) {
-                    return { ...todo, completed: !todo.completed} //"..." znamená id: todo.id, text: todo.text, ...
-                }
-                return todo
-            })
-            console.log(todos)
-            renderTodos();
-        })
-
-        const deleteButton = document.createElement("button")
-        deleteButton.innerHTML = 'delet'
-
-        deleteButton.addEventListener('click', () => {
-            todos = todos.filter((todo) => todo.id !== item.id)
-            renderTodos();
-        })
-
-        todoItemElement.appendChild(todoSpanElement)
-        todoItemElement.appendChild(deleteButton)
-
-        todoListElement?.appendChild(todoItemElement)
-
-    })
-}
